Split file into test+generated code. Add comment
